requirements.txt
ccxt>=4.2.15
pymongo>=4.6.1
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.3
tweepy>=4.14.0
newsapi-python>=0.2.7
ta>=0.11.0  # Technical Analysis library
matplotlib>=3.8.2
seaborn>=0.13.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
pydantic>=2.6.0 